HtmlBridge Phase 3 ratchet: de-list DomBridge.cs facade (split window-load + parsing partials)
Shrink the primary facade DomBridge.cs (1013 lines) toward the exit-criterion
500-800-line target by peeling two cohesive behaviour clusters — neither of
which is composition/compat facade material — into sibling partials:

- DomBridge.WindowLoad.cs (209 lines): the window `load` lifecycle and
  window-target event dispatch (FireWindowLoadEvent, both DispatchWindowEvent
  overloads, BuildWindowFramesArray / CollectWindowFrames).
- DomBridge.HtmlParsing.cs (159 lines): initial HTML/doctype ingestion and
  inline-style parsing (ParseHtml, ParseStyle, IsAcceptableInlineValue, the
  DocTypePattern field and its [GeneratedRegex]).

The now-unused System.Text.RegularExpressions using is dropped from the facade.

Pure partial-class relocation — no signature, accessibility, or logic change,
so behaviour-identical by construction. DomBridge.cs drops 1013 -> 682, now
inside the 500-800 facade band, so its OversizedFileExemptions entry is removed
and the stale-exemption check enforces it stays under the limit. Debt list:
six -> five files.

Broiler.HtmlBridge.Dom builds clean; HtmlBridgeArchitectureGuardTests 14/14
green (0 offenders / 0 stale exemptions); load-event / parsing / inline-style
coverage (Acid3RegressionTests, Acid3SvgAndParsingRegressionTests,
Acid3HtmlElementRegressionTests, DoctypeSentinelMigrationTests,
InlineStyleDropDiagnosticsTests, InlineStyleWriteThroughTests, 90 tests with
the guard) green.
