+++ src/Broiler.HtmlBridge.Dom/DomBridge.WindowLoad.cs
@@ -1,0 +1,210 @@
+using Broiler.JavaScript.BuiltIns.Null;
+using Broiler.JavaScript.BuiltIns.Boolean;
+using Broiler.JavaScript.BuiltIns.Array;
+using Broiler.JavaScript.BuiltIns.String;
+using Broiler.JavaScript.BuiltIns.Number;
+using Broiler.JavaScript.BuiltIns.Function;
+using Broiler.JavaScript.Runtime;
+using Broiler.JavaScript.Storage;
+using Broiler.JavaScript.Engine;
+using Broiler.HtmlBridge.Logging;
+using Broiler.Dom;
+
+namespace Broiler.HtmlBridge;
+
+/// <summary>
+/// Sibling partial peeled out of <c>DomBridge.cs</c> (Phase 3 ratchet, 2026-07-17) to keep the
+/// facade under the 750-line guard: the window <c>load</c> lifecycle and window-target event
+/// dispatch. Fires <c>window.onload</c> / bare <c>onload</c>, the window <c>load</c> listeners,
+/// and the <c>&lt;body&gt;</c> load event, and builds the <c>window.frames</c> array from the
+/// document's same-origin iframes. Pure partial-class relocation — no signature, accessibility,
+/// or logic change.
+/// </summary>
+public sealed partial class DomBridge
+{
+    /// <summary>
+    /// Fires the <c>load</c> event on the <c>&lt;body&gt;</c> element, which
+    /// triggers the inline <c>onload</c> attribute handler as well as any
+    /// <c>addEventListener('load', …)</c> listeners registered on the body.
+    /// In browsers, the body's <c>onload</c> fires after all synchronous
+    /// scripts have executed. This is critical for test harnesses like Acid3,
+    /// which use <c>&lt;body onload="update()"&gt;</c> to bootstrap the
+    /// test runner.
+    /// </summary>
+    public void FireWindowLoadEvent()
+    {
+        ThrowIfDisposed();
+        if (_jsContext == null) return;
+
+        _jsContext["frames"] = BuildWindowFramesArray();
+
+        var htmlEl = Elements.FirstOrDefault(e =>
+            string.Equals(e.TagName, "html", StringComparison.OrdinalIgnoreCase));
+        if (htmlEl != null)
+            FireDescendantOnloads(htmlEl);
+
+        // 1. Fire window.onload if it was set by script.
+        //    In browsers, setting `window.onload = fn` registers a handler
+        //    that fires when the page finishes loading.  This is distinct
+        //    from the <body onload="…"> inline attribute handler.
+        try
+        {
+            _jsContext.Eval(@"
+(function() {
+  // A page may register the load handler either as `window.onload = fn`
+  // or as a bare `onload = fn` assignment. In a browser `window` IS the
+  // global object so both are the same property; in this engine the global
+  // object and `window` are distinct, so a bare `onload = fn` lands on the
+  // global (globalThis.onload) and never on window.onload. Check both, with
+  // window.onload winning when present.
+  var h = null;
+  if (typeof window.onload === 'function') h = window.onload;
+  else if (typeof onload === 'function') h = onload;
+  if (h) {
+    try { h(); } catch(e) {}
+  }
+})();");
+        }
+        catch (Exception ex)
+        {
+            RenderLogger.LogError(LogCategory.JavaScript, "DomBridge.FireWindowLoadEvent",
+                $"Error firing window.onload: {ex.Message}", ex);
+        }
+
+        try
+        {
+            DispatchWindowEvent("load");
+        }
+        catch (Exception ex)
+        {
+            RenderLogger.LogError(LogCategory.JavaScript, "DomBridge.FireWindowLoadEvent",
+                $"Error firing window load listeners: {ex.Message}", ex);
+        }
+
+        // 2. Fire <body onload="…"> attribute handler and any load event
+        //    listeners registered on the body element.
+        // Find the <body> element by traversing the document tree.
+        // The body is a child of <html> (documentElement), which is a
+        // child of the document node.
+        DomElement? body = null;
+        if (htmlEl != null)
+        {
+            body = ChildElements(htmlEl).FirstOrDefault(c =>
+                string.Equals(c.TagName, "body", StringComparison.OrdinalIgnoreCase));
+        }
+        if (body == null) return;
+
+        // Ensure the body's JS object is created so inline event attributes are compiled
+        ToJSObject(body);
+
+        // Dispatch a 'load' event on the body element. This covers inline
+        // attributes, property-assigned handlers (document.body.onload = fn),
+        // and addEventListener registrations using the same event path.
+        try
+        {
+            if (_jsContext.Eval("(function() { var e = document.createEvent('Event'); e.initEvent('load', false, false); return e; })()") is JSObject evt)
+                DispatchEventOnElement(body, evt);
+        }
+        catch (Exception ex)
+        {
+            RenderLogger.LogError(LogCategory.JavaScript, "DomBridge.FireWindowLoadEvent",
+                $"Error firing window load event: {ex.Message}", ex);
+        }
+    }
+
+    private JSBoolean DispatchWindowEvent(string eventType, bool bubbles = false)
+    {
+        var evt = new JSObject();
+        evt.FastAddValue((KeyString)"type", new JSString(eventType), JSPropertyAttributes.EnumerableConfigurableValue);
+        evt.FastAddValue((KeyString)"bubbles", bubbles ? JSBoolean.True : JSBoolean.False, JSPropertyAttributes.EnumerableConfigurableValue);
+        return DispatchWindowEvent(evt);
+    }
+
+    private JSBoolean DispatchWindowEvent(JSObject evt)
+    {
+        if (_jsContext == null || _windowJSObject == null)
+            return JSBoolean.True;
+
+        var eventType = evt[(KeyString)"type"]?.ToString() ?? "unknown";
+        evt.FastAddValue((KeyString)"target", _windowJSObject, JSPropertyAttributes.EnumerableConfigurableValue);
+        evt[(KeyString)"srcElement"] = _windowJSObject;
+        evt.FastAddValue((KeyString)"currentTarget", _windowJSObject, JSPropertyAttributes.EnumerableConfigurableValue);
+        evt.FastAddValue((KeyString)"eventPhase", new JSNumber(2), JSPropertyAttributes.EnumerableConfigurableValue);
+
+        var immediateStopped = false;
+        var prevented = evt[(KeyString)"defaultPrevented"] is JSValue defaultPreventedValue &&
+                        defaultPreventedValue.BooleanValue;
+        var currentListenerPassive = false;
+        var legacyCancelBubble = false;
+        evt[(KeyString)"defaultPrevented"] = prevented ? JSBoolean.True : JSBoolean.False;
+        evt.FastAddValue((KeyString)"stopPropagation",
+            new JSFunction((in _) => JsCallbackStopPropagation001Core(ref legacyCancelBubble, in _), "stopPropagation", 0),
+            JSPropertyAttributes.EnumerableConfigurableValue);
+        evt.FastAddValue((KeyString)"stopImmediatePropagation",
+            new JSFunction((in _) => JsCallbackStopImmediatePropagation002Core(ref immediateStopped, ref legacyCancelBubble, in _), "stopImmediatePropagation", 0),
+            JSPropertyAttributes.EnumerableConfigurableValue);
+        evt.FastAddValue((KeyString)"preventDefault",
+            new JSFunction((in _) => JsCallbackPreventDefault003Core(currentListenerPassive, evt, ref prevented, in _), "preventDefault", 0),
+            JSPropertyAttributes.EnumerableConfigurableValue);
+        evt.FastAddProperty(
+            (KeyString)"cancelBubble",
+            new JSFunction((in _) => legacyCancelBubble ? JSBoolean.True : JSBoolean.False, "get cancelBubble"),
+            new JSFunction((in setArgs) => JsCallbackSetCancelBubble005Core(ref legacyCancelBubble, in setArgs), "set cancelBubble"),
+            JSPropertyAttributes.EnumerableConfigurableProperty);
+        evt.FastAddProperty(
+            (KeyString)"returnValue",
+            new JSFunction((in _) => prevented ? JSBoolean.False : JSBoolean.True, "get returnValue"),
+            new JSFunction((in setArgs) => JsCallbackSetReturnValue007Core(currentListenerPassive, evt, ref prevented, in setArgs), "set returnValue"),
+            JSPropertyAttributes.EnumerableConfigurableProperty);
+        evt.FastAddValue((KeyString)"composedPath",
+            new JSFunction((in _) => new JSArray(_windowJSObject), "composedPath", 0),
+            JSPropertyAttributes.EnumerableConfigurableValue);
+
+        if (_eventTargets.TryGetWindowListeners(eventType, out var listeners))
+        {
+            foreach (var registration in listeners.ToList())
+            {
+                if (immediateStopped)
+                    break;
+
+                currentListenerPassive = registration.Passive;
+                InvokeEventListener(registration.Listener, evt, "DomBridge.window.dispatchEvent");
+                currentListenerPassive = false;
+
+                if (registration.Once)
+                    listeners.Remove(registration);
+            }
+        }
+
+        evt[(KeyString)"currentTarget"] = JSNull.Value;
+        evt[(KeyString)"eventPhase"] = new JSNumber(0);
+        return prevented ? JSBoolean.False : JSBoolean.True;
+    }
+
+    private JSArray BuildWindowFramesArray()
+    {
+        var frames = new List<JSValue>();
+        CollectWindowFrames(DocumentElement, frames);
+        return new JSArray([.. frames]);
+    }
+
+    private void CollectWindowFrames(DomElement element, List<JSValue> frames)
+    {
+        foreach (var child in ChildElements(element))
+        {
+            if (IsText(child))
+                continue;
+
+            if (string.Equals(child.TagName, "iframe", StringComparison.OrdinalIgnoreCase))
+            {
+                var src = TryGetAttribute(child, "src", out var srcValue) ? srcValue : string.Empty;
+                if (!IsCrossOrigin(src, _pageUrl))
+                    frames.Add(_subWindows.GetOrCreate(child));
+            }
+
+            // Sub-documents are severed (P4.4b) — never in-tree children — so the walk always
+            // descends normal element children.
+            CollectWindowFrames(child, frames);
+        }
+    }
+}
